Work around a change in (presumably) binutils on Fedora 28.
Without this, symbols used in dynamically loaded modules are looked up
in all .so files that have been loaded so far, and in the module
itself last.  It so happens that pyapi defines and uses a symbol
string_copy that is then found in the wrong library.  With this, the
lookup happens in the module itself first, so the correct definition
is found.  The difference with before seems to be that the symbol is
looked up at all.