On Unix/Linux don't link dynamic libraries from the server into modules.
On Unix/Linux, when linking, references to the dynamic libraries on
the link command line are filled in so that when the module is loaded
there are actually two copies of the linked-to library loaded: the one
the server was linked to and the one the dynamically loaded module was
linked to.
On Windows, the linker merely adds stubs for references and when the
module is loaded those stubs find the actual code in the already
loaded dynamic libraries.