Add new table in the vault
In SQL the new table definition would be:

CREATE TABLE remote_table_credentials (
  uri         string,
  local_user  string,
  remote_user string,
  remote_pass string,
  PRIMARY KEY (uri, local_user)
);

The idea is that access to the remote user name and password will need
to be controlled. When trying to access the remote user and password
the vault will need to check if the current user is an admin or
matches the local user that has been stored.